requirements.txt
# Core AI/ML Libraries
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
datasets>=2.12.0
accelerate>=0.20.0
peft>=0.4.0

# Vector Databases
chromadb>=0.4.0
faiss-cpu>=1.7.4
qdrant-client>=1.6.0

# RAG and Document Processing
langchain>=0.0.200
langchain-community>=0.0.20
pypdf>=3.0.0
python-docx>=0.8.11
python-multipart>=0.0.6

# API and Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
requests>=2.31.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Monitoring and Logging
prometheus-client>=0.17.0
psutil>=5.9.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0
